chore: consolidate type files
